IMPALA-4803: add 2.8 change log
The ASF Jira site provides a mechanism to render so-called "release
notes" in HTML. The rendered HTML is more like what I'd call a change
log, so I try to use that phrase here and on the site. The rendered HTML
shows resolved Jira issues for a given release, grouped by issueType.
Curating the HTML requires

1. Reordering the groups: the Jira release notes lists Sub-tasks first,
which are less interesting than New Features, Epics, or Improvements. I
manually reordered the groups to have the more interesting groups first,
and less interesting groups later.

2. Adding some preamble HTML to make this a distinct page.

Change-Id: Icaaa6ee1182702b2b5986ff44041fcf40fe7af0e